ARTIFACT/SAFE/requirements.txt
pydantic==2.8.2
pyyaml==6.0.2
pandas==2.2.2

langchain==0.2.14
langchain-core==0.2.39
langchain-openai==0.1.22
langchain-anthropic==0.1.23

langgraph==0.2.35
